+++ src/main/java/ru/mosquiito/services/accountConfirmation/AccountConfirmationDataAccess.java
@@ -4,5 +4,5 @@ import ru.mosquiito.domain.AccountConfirmation;
 
 public interface AccountConfirmationDataAccess {
     AccountConfirmation findByCode(String code);
-    AccountConfirmation findById(Integer id);
+    AccountConfirmation findById(Long id);
 }
