+++ src/main/java/ru/mosquiito/services/accountConfirmation/AccountDataAccess.java
@@ -3,6 +3,6 @@ package ru.mosquiito.services.accountConfirmation;
 import ru.mosquiito.domain.Account;
 
 public interface AccountDataAccess {
-    Account findById(Integer id);
+    Account findById(Long id);
     Account update(Account account);
 }
